Refactor fbscraper.py to remove facepy dependency
facepy's fork that has been in active use in naarad has not been updated
in a while, and falls way behind vanilla Graph API in terms of
functionality as well as documentation. Directly interfacing with the
Graph API makes it easier to work with the API, and helps keeping the
code simple enough for future contributors by not adding another
abstraction layer in the middle.

Notes:

  - Regenerated requirements.txt using 'pipreqs' (Please do check if the
    list is exhaustive)